Exercise 4.21

+++ controllers/blogs.js
@@ -18,7 +18,6 @@ blogsRouter.post('/', async (request, response) => {
     return response.status(400).json({ error: 'title or url missing' })
   }
 
-  console.log(request.token)
   if (!request.token) {
     return response.status(401).send({error: 'authorization token missing'})
   }
@@ -58,6 +57,21 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  if (!request.token) {
+    return response.status(401).send({error: 'authorization token missing'})
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(400).json({ error: 'unauthorized action' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
